move extension helper file to main one
